fix up code
