--- src/ExportPro.Common/ExportPro.Common.Shared/Library/ApiServiceResponse.cs
@@ -1,13 +1,0 @@
-﻿using System.Net;
-
-namespace ExportPro.Common.Shared.Library
-{
-    public class ApiServiceResponse
-    {
-        public string? Message { get; set; }
-        public string DetailsMessage { get; set; }
-        public HttpStatusCode ApiState { get; set; }
-        public string? ErrorCode { get; set; }
-        public List<string> ValidationErrors { get; set; }
-    }
-}
