--- src/ExportPro.Common/ExportPro.Common.Shared/Library/BadRequestApiServiceResponse.cs
@@ -1,16 +1,0 @@
-﻿using System.Net;
-
-
-namespace ExportPro.Common.Shared.Library
-{
-    public class BadRequestApiServiceResponse : ApiServiceResponse
-    {
-        public BadRequestApiServiceResponse(string message = null, string errorCode = null, List<string> validationErrors = null)
-        {
-            ApiState = HttpStatusCode.BadRequest;
-            Message = message;
-            ValidationErrors = validationErrors;
-            ErrorCode = errorCode ?? HttpStatusCode.BadRequest.ToString();
-        }
-    }
-}
