+++ src/ExportPro.Common/ExportPro.Common.Shared/Library/BadRequestResponse.cs
@@ -1,0 +1,14 @@
+﻿using System.Net;
+
+
+namespace ExportPro.Common.Shared.Library
+{
+    public class BadRequestResponse : BaseResponse
+    {
+        public BadRequestResponse(List<string> messages = null)
+        {
+            ApiState = HttpStatusCode.BadRequest;
+            Messages = messages;
+        }
+    }
+}
