+++ src/ExportPro.Common/ExportPro.Common.Shared/Library/BaseResponseT.cs
@@ -1,0 +1,7 @@
+﻿namespace ExportPro.Common.Shared.Library;
+
+public class BaseResponseT<T> : BaseResponse
+{
+     public T? Data { get; set; }
+}
+
