--- src/ExportPro.Common/ExportPro.Common.Shared/Library/NotFoundApiServiceResponse.cs
@@ -1,15 +1,0 @@
-﻿using System.Net;
-
-
-namespace ExportPro.Common.Shared.Library
-{
-    public class NotFoundApiServiceResponse<T> : ApiServiceResponse<T>
-    {
-        public NotFoundApiServiceResponse(string message = null, string errorCode = null)
-        {
-            ApiState = HttpStatusCode.NotFound;
-            Message = message;
-            ErrorCode = errorCode ?? HttpStatusCode.NotFound.ToString();
-        }
-    }
-}
