--- src/ExportPro.Common/ExportPro.Common.Shared/Library/SuccessApiServiceResponse.cs
@@ -1,15 +1,0 @@
-﻿
-using System.Net;
-
-
-namespace ExportPro.Common.Shared.Library
-{
-    public class SuccessApiResponse : ApiServiceResponse
-    {
-        public SuccessApiResponse(string message = null)
-        {
-            ApiState = HttpStatusCode.OK;
-            Message = message;
-        }
-    }
-}
