--- src/ExportPro.Common/ExportPro.Common.Shared/Library/SuccessApiServiceResponseT.cs
@@ -1,15 +1,0 @@
-﻿using System.Net;
-
-
-namespace ExportPro.Common.Shared.Library
-{
-    public class SuccessApiServiceResponse<T> : ApiServiceResponse<T>
-    {
-        public SuccessApiServiceResponse(T data, string message = null)
-        {
-            Data = data;
-            ApiState = HttpStatusCode.OK;
-            Message = message;
-        }
-    }
-}
