+++ src/ExportPro.Common/ExportPro.Common.Shared/Library/SuccessResponse.cs
@@ -1,0 +1,15 @@
+﻿
+using System.Net;
+
+
+namespace ExportPro.Common.Shared.Library;
+
+public class SuccessApiResponse : BaseResponse
+{
+    public SuccessApiResponse(string message = null)
+    {
+        ApiState = HttpStatusCode.Accepted;
+        Messages = [message];
+    }
+}
+
