+++ src/ExportPro.Common/ExportPro.Common.Shared/Library/SuccessResponseT.cs
@@ -1,0 +1,15 @@
+﻿using System.Net;
+
+
+namespace ExportPro.Common.Shared.Library;
+
+public class SuccessApiServiceResponse<T> : BaseResponseT<T>
+{
+    public SuccessApiServiceResponse(T data, string message = null)
+    {
+       Data = data;
+       ApiState = HttpStatusCode.OK;
+       Messages = [message];
+    }
+}
+
